\name{freq.hap}
\docType{data}
\alias{freq.hap}
\alias{simuBkMap}
\alias{step3way}
\title{Examples for Haplotype Frequencies for Simulation}
\description{
Both \code{freq.hap} and \code{simuBkMap} are \code{data.frame}s containing haplotype information, including the
haplotype block identifier, haplotype, and haplotype frequency. While \code{freq.hap} is a data frame consisting of
20 rows and 3 columns, \code{simuBkMap} consists of 66 rows and 3 columns.

\code{step3way} is a list internally used for simulation, containing some indexes and sampling frequencies.  
}
\author{Qing Li, mail2qing@yahoo.com}
\keyword{datasets}

