\name{LDdata}
\alias{LDdata}
\docType{data}
\title{
Simulated LD Block Data
}
\description{
A 500 x 50 matrix containing genotype data for 10 LD blocks each consisting of 5 SNPs in strong LD.
}

\author{Holger Schwender, \email{holger.schwender@udo.edu}}

\seealso{\code{\link{getLD}}, \code{\link{findLDblocks}}}

\keyword{datasets}
