\name{read.pedfile}
\alias{read.pedfile}

\title{
Reading a Ped File
}
\description{
Reads a ped file into R and creates a data frame or data table in ped format, or transform the ped file into a matrix in genotype format.
}
\usage{
read.pedfile(file, first.row = NA, coded = NULL, naVal = 0, sep = " ", 
   p2g = FALSE, non.rs.IDs = FALSE, asDataTable = FALSE, cols4ID=FALSE,
   ref = NULL, alt = NULL)
}

\arguments{
  \item{file}{the filename (if necessary with path) of a ped file that should be read into R.
}
  \item{first.row}{logical indicating whether the first row of \code{file} also contains data for a subject. If \code{FALSE}, the first row is assumed to contain the SNP names. By default, \code{read.pedfile} tries to figure out automatically if the first column contains the SNP names or data for a subject.
}
  \item{coded}{a character string stating how the alleles of the SNPs are coded. Possible values are \code{"12"}, \code{"AB"}, \code{"1234"}, \code{"ATCG"}. For details, see \code{\link{ped2geno}}. By default, \code{read.pedfile} tries to figure out automatically how the alleles are coded. 
}
  \item{naVal}{value or character string specifying how missing values in the SNP data are coded.
}
  \item{sep}{character string specifying how the SNP names in the first row of \code{file} are separated. Ignored if \code{first.row = TRUE}.
}
  \item{p2g}{logical indicating whether the ped file should be transformed into a matrix in genotype format. If \code{FALSE}, depending on the specification of \code{asDataTable}, either a data frame or data table in ped format is returned. Otherwise, \code{\link{ped2geno}} is called within \code{read.pedfile} to transform the data table into a matrix in genotype format, and the matrix is returned.
}
  \item{non.rs.IDs}{logical indicating whether (some of) the SNP names are specified by other names than rs-IDs.
}
  \item{asDataTable}{logical indication whether a data table or a data frame should be returned, if \code{p2g} is set to \code{FALSE}. By default, a data frame is returned in the case of \code{p2g = FALSE}.
}
  \item{cols4ID}{logical indicating whether columns should be added to output matrix containing the family ID and the individual ID. If \code{FALSE}, the individual IDs are used as the row names of the output matrix.}

\item{ref}{character vector specifying the reference alleles of the SNPs in the ped file, where the alleles must be coded by \code{"A"}, \code{"T"}, \code{"C"}, and \code{"G"}. Missing values are not allowed. If specified, this vector is added as attribute to the genotype matrix returned by \code{read.pedfile} if \code{p2g = TRUE}. Ignored if \code{p2g = FALSE}. Must be specified, if \code{alt} is specified.} 

\item{alt}{character vector specifying the alternative alleles of the SNPs in the ped file, where the alleles must be coded by \code{"A"}, \code{"T"}, \code{"C"}, and \code{"G"}. It is assumed that the SNPs in the ped file are biallelic so that each SNP has one alternative allele. Missing values are not allowed. If specified, this vector is added as attribute to the genotype matrix returned by \code{read.pedfile} if \code{p2g = TRUE}. Ignored if \code{p2g = FALSE}. Must be specified, if \code{alt} is specified.} 

}

\value{If \code{p2g = FALSE}, depending on the specification of \code{asDataTable}, either a data frame or data table in ped format. If \code{p2g = TRUE}, a matrix in genotype format is returned.
}

\author{
Holger Schwender, \email{holger.schwender@udo.edu}
}

\seealso{
\code{\link{ped2geno}}
}

\examples{\dontrun{
# Assuming there is a ped-file called pedfile.ped in the 
# R working directory, this file can be read into R by
ped <- read.pedfile("pedfile.ped")

# The resulting data frame is in the typical ped format
# which needs to be transformed into the genotype format
# for applications of most of the functions in the trio
# package. This transformation can be done by
geno <- ped2geno(ped)

# This transformation can also be done directly when
# reading the ped-file into R by
geno2 <- read.pedfile("pedfile.ped", p2g = TRUE)
}}

\keyword{IO}
\keyword{file}

