\name{simuBkMap}
\docType{data}
\alias{simuBkMap}
\title{Haplotype frequencies for simulation}
\description{
It is a data.frame containing haplotype information, including the
haplotype block identifier, haplotype, and haplotype frequency.  
}
\usage{simuBkMap}
\format{A data.frame of 66 rows and 3 columns.}
\keyword{datasets}

