\name{trio.gen.err}
\docType{data}
\alias{trio.gen.err}
\title{Genotype data for trios with Mendelian errors}
\description{
This data set is a matrix containing genotype data for 10 SNPs in 100
trios. The first two columns contain family identifier and individual
identifier. In the columns afterwards, each SNPs is encoded in one
variable denoting the number of minor alleles. Different from trio.gen1,
this data set contains Mendelian errors.
}
\usage{trio.gen.err}
\format{A matrix with 12 columns and 300 rows.}
\keyword{datasets}

