\name{trio.gen2}
\docType{data}
\alias{trio.gen2}
\title{Genotype data for trios with missing genotypes}
\description{
This data set is a matrix containing genotype data for 10 SNPs in 100
trios. The first two columns contain family identifier and individual
identifier. In the columns afterwards, each SNPs is encoded in one
variable denoting the number of minor alleles. Different from trio.gen1,
this data set contain missing genotypes, denoted as integer 0.
}
\usage{trio.gen2}
\format{A matrix with 12 columns and 300 rows.}
\keyword{datasets}

