\name{trio.ped2}
\docType{data}
\alias{trio.ped2}
\title{Genotype data for pedigrees in the format of a linkage file, with
missing genotypes}
\description{
This data set is a matrix containing genotype data for 10 SNPs in 100
pedigrees. The first six columns identify the family
structure of the data, and the phenotype. Besides the variables
providing information on the family structure and the phenotypes
(columns 1 to 6), each SNPs is encoded in two variables denoting the alleles.
Different from data set trio.ped1, trio.ped2 contain missing genotypes,
denoted as interger 0.
}
\usage{trio.ped2}
\format{A matrix with 26 columns and 300 rows.}
\keyword{datasets}

