\name{freq.hap}
\docType{data}
\alias{freq.hap}
\title{Haplotype frequencies}
\description{
It is a data.frame containing haplotype information, including the
haplotype block identifier, haplotype, and haplotype frequency.  
}
\usage{freq.hap}
\format{A data.frame of 20 rows and 3 columns.}
\keyword{datasets}

