\name{trio.gen1}
\docType{data}
\alias{trio.gen1}
\title{Genotype data for trios}
\description{
This data set is a matrix containing genotype data for 10 SNPs in 100
trios. The first two columns contain family identifier and individual
identifier. In the columns afterwards, each SNPs is encoded in one variable denoting the number of minor alleles.
}
\usage{trio.gen1}
\format{A matrix with 12 columns and 300 rows.}
\keyword{datasets}

